% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monkeylearn-defunct.R
\name{monkeylearn-defunct}
\alias{monkeylearn-defunct}
\alias{monkeylearn_extract}
\title{Defunct functions in monkeylearn}
\usage{
monkeylearn_extract(...)
}
\description{
These functions are gone, no longer available.
}
\details{
\itemize{
 \item \code{\link{monkeylearn_extract}}: Use monkey_extract() instead.
 \item \code{\link{monkeylearn_classify}}: Use monkey_classify() instead.
}
}
