% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monkeylearn-package.R
\docType{package}
\name{monkeylearn-package}
\alias{monkeylearn}
\alias{monkeylearn-package}
\title{monkeylearn: Accesses the Monkeylearn API for Text Classifiers and Extractors}
\description{
Allows using some services of Monkeylearn <http://monkeylearn.com/> which is
    a Machine Learning platform on the cloud for text analysis (classification and extraction).
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/ropensci/monkeylearn}
  \item \url{http://docs.ropensci.org/monkeylearn/}
  \item Report bugs at \url{http://github.com/ropensci/monkeylearn/issues}
}

}
\author{
\strong{Maintainer}: Maëlle Salmon \email{maelle.salmon@yahoo.se} (\href{https://orcid.org/0000-0002-2815-0399}{ORCID})

Authors:
\itemize{
  \item Amanda Dobbyn \email{amanda.e.dobbyn@gmail.com}
}

Other contributors:
\itemize{
  \item Thomas Leeper (Thomas Leeper reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/45) [reviewer]
  \item Jeroen Ooms [contributor]
  \item rOpenSci (https://ropensci.org/) [funder]
  \item Earlybird Software (http://earlybird.co/) [funder]
}

}
\keyword{internal}
