% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifiers.R
\name{monkeylearn_classifiers}
\alias{monkeylearn_classifiers}
\title{monkeylearn_classifiers}
\usage{
monkeylearn_classifiers(private = FALSE)
}
\arguments{
\item{private}{default is FALSE, whether to show private modules instead of public modules}
}
\value{
A data.frame, but nothing now, in dev.
}
\description{
List of Monkeylearn classifiers modules
}

