% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor.R
\name{monkeylearn_extract}
\alias{monkeylearn_extract}
\title{monkeylearn_extract}
\usage{
monkeylearn_extract(request, key = monkeylearn_key(quiet = TRUE),
  extractor_id = "ex_isnnZRbS")
}
\arguments{
\item{request}{A vector of characters}

\item{key}{The API key}

\item{extractor_id}{The ID of the extractor}
}
\value{
A list of two data.frames (dplyr tbl_df), one with the results, the other with headers including the number of remaining queries as "x.query.limit.remaining".
}
\description{
Access to Monkeylearn extractors modules
}
\details{
Find IDs of extractors using \url{https://app.monkeylearn.com/main/explore}.
}
\examples{
text <- "In the 19th century, the major European powers had gone to great lengths to maintain a balance of power throughout Europe, resulting in the existence of a complex network of political and military alliances throughout the continent by 1900.[7] These had started in 1815, with the Holy Alliance between Prussia, Russia, and Austria. Then, in October 1873, German Chancellor Otto von Bismarck negotiated the League of the Three Emperors (German: Dreikaiserbund) between the monarchs of Austria-Hungary, Russia and Germany."
output <- monkeylearn_extract(request = text)
output
}

