% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{monkeylearn_key}
\alias{monkeylearn_key}
\title{Retrieve Monkeylearn API key}
\usage{
monkeylearn_key(quiet = TRUE)
}
\value{
An Monkeylearn API Key
}
\description{
Retrieve Monkeylearn API key
}
\details{
Looks in env var \code{MONKEYLEARN_KEY}
}
\keyword{internal}

