% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{monkeylearn_classify}
\alias{monkeylearn_classify}
\title{monkeylearn_classify}
\usage{
monkeylearn_classify(request, key = monkeylearn_key(quiet = TRUE))
}
\arguments{
\item{request}{A vector of characters}

\item{key}{The API key}
}
\description{
monkeylearn_classify
}
\examples{
text1 <- "lions are very big animals"
text2 <- "i want to buy an iphone"
request <- c(text1, text2)
output <- monkeylearn_classify(request)
output
}

