% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{monkeylearn_classify}
\alias{monkeylearn_classify}
\title{monkeylearn_classify}
\usage{
monkeylearn_classify(request, key = monkeylearn_key(quiet = TRUE))
}
\arguments{
\item{request}{}

\item{key}{}
}
\description{
monkeylearn_classify
}
\examples{
text1 <- "lions are very big animals"
text2 <- "i want to buy an iphone"
request <- toJSON(list(text_list = c(text1, text2)))
result <- monkeylearn_classify(request)
httr::content(result)
}

