% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor.R
\name{monkeylearn_extract}
\alias{monkeylearn_extract}
\title{monkeylearn_extract}
\usage{
monkeylearn_extract(request, key = monkeylearn_key(quiet = TRUE),
  extractor_id = "ex_isnnZRbS")
}
\arguments{
\item{request}{A vector of characters (each text smaller than 50kB)}

\item{key}{The API key}

\item{extractor_id}{The ID of the extractor}
}
\value{
A list of two data.frames (dplyr tbl_df), one with the results, the other with headers including the number of remaining queries as "x.query.limit.remaining".
}
\description{
Access to Monkeylearn extractors modules
}
\details{
Find IDs of extractors using \url{https://app.monkeylearn.com/main/explore}.
Within the free plan, you can make up to 20 requests per minute.
 You can use batch to send up to 200 texts to be analyzed within the API
 (classification or extraction) with each request.
So for example, if you need to analyze 6000 tweets,
instead of doing 6000 requests to the API, you can use batch to send 30 requests,
each request with 200 tweets.
The function automatically makes these batch calls and waits if there is a throttle limit error,
but you might want to control the process yourself using several calls to the function.
}
\examples{
text <- "In the 19th century, the major European powers had gone to great lengths to maintain a balance of power throughout Europe, resulting in the existence of a complex network of political and military alliances throughout the continent by 1900.[7] These had started in 1815, with the Holy Alliance between Prussia, Russia, and Austria. Then, in October 1873, German Chancellor Otto von Bismarck negotiated the League of the Three Emperors (German: Dreikaiserbund) between the monarchs of Austria-Hungary, Russia and Germany."
output <- monkeylearn_extract(request = text)
output
}

